refactor: move all utils used on calendar to utils folder
