+++ json-server/db.json
@@ -11,6 +11,12 @@
       "dateStart": "2023-07-18T16:39:10.186Z",
       "dateEnd": "2023-07-18T16:39:15.600Z",
       "id": 2
+    },
+    {
+      "title": "No name",
+      "dateStart": "2023-08-03T14:25:08.039Z",
+      "dateEnd": "2023-08-03T14:25:13.079Z",
+      "id": 3
     }
   ]
 }
