--- src/lib/utils.ts
@@ -1,1 +1,0 @@
-export const addZero = (num: number) => (num < 10 ? `0${num}` : `${num}`);
